refactor: 새로운 파티 관련 컴포넌트 정리
